pyproject.toml
[tool.poetry]
name = "intro-llama2"
version = "0.1.0"
description = ""
authors = ["Helton Carlos de Souza <heltoncarlossouza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "0.0.335"
tiktoken = "^0.5.1"
pypdf = "^3.17.0"
gpt4all = "^2.0.2"
redis = "^5.0.1"
openai = "^1.2.3"
chromadb = "^0.4.17"
faiss-cpu = "^1.7.4"
streamlit = "^1.28.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
